+++ web/ui/react-app/src/pages/Config.tsx
@@ -1,9 +1,9 @@
-import React, {FC, useState} from 'react';
-import {Button} from 'reactstrap';
-import {RouteComponentProps} from '@reach/router';
+import React, { FC, useState } from 'react';
+import { Button } from 'reactstrap';
+import { RouteComponentProps } from '@reach/router';
 import CopyToClipboard from 'react-copy-to-clipboard';
-import {withStatusIndicator} from '../withStatusIndicator';
-import {useFetch} from '../utils/useFetch';
+import { withStatusIndicator } from '../withStatusIndicator';
+import { useFetch } from '../utils/useFetch';
 
 import './Config.css';
 
@@ -14,12 +14,12 @@ interface ConfigContentProps {
   data?: YamlConfig;
 }
 
-const YamlContent = ({yaml}: YamlConfig) => <pre className="config-yaml">{yaml}</pre>;
+const YamlContent = ({ yaml }: YamlConfig) => <pre className="config-yaml">{yaml}</pre>;
 YamlContent.displayName = 'Config';
 
 const ConfigWithStatusIndicator = withStatusIndicator(YamlContent);
 
-export const ConfigContent: FC<ConfigContentProps> = ({error, data}) => {
+export const ConfigContent: FC<ConfigContentProps> = ({ error, data }) => {
   const [copied, setCopied] = useState(false);
   const config = data && data.yaml;
   return (
@@ -40,12 +40,12 @@ export const ConfigContent: FC<ConfigContentProps> = ({error, data}) => {
       </h2>
       <ConfigWithStatusIndicator error={error} isLoading={!config} yaml={config} />
     </>
-  )
+  );
 };
 
 const Config: FC<RouteComponentProps> = () => {
-  const {response, error} = useFetch<YamlConfig>(`/api/v1/status/config`);
-  return <ConfigContent error={error} data={response.data}/>;
+  const { templateConfigResp, error } = useFetch<YamlConfig>(`/api/v1/status/config`);
+  return <ConfigContent error={error} data={templateConfigResp.data} />;
 };
 
 export default Config;