+++ web/ui/react-app/src/pages/index.ts
@@ -1,5 +1,6 @@
 import Config from './Config';
 import Flags from './Flags';
 import Status from './Status';
+import Playground from './Playground';
 
-export {Config, Status, Flags};
+export { Config, Status, Flags, Playground };